Extract shared float-domain bicubic resize to image_resample_float.h
Move BicubicResizeU8ToFloatRGB and detail:: helpers (bicubic_kernel,
precompute_coeffs, resample_horizontal/vertical) from
image_transforms_gemma4.hpp into a new shared header
image_resample_float.h.

This makes the float-domain resize reusable by other image processors
(phi3, phi4, mllama, etc.) if they need to match torchvision behavior
precisely, as suggested in review.

No functional change — image_transforms_gemma4.hpp now includes the
shared header instead of inlining the code.
